+++ eras/conway/impl/CHANGELOG.md
@@ -52,6 +52,9 @@
 
 ### `testlib`
 
+* Add `getConstitution` to Conway ImpTest
+* Change return type of `setupSingleDRep` to Credential instead of KeyHash
+* Add `registerInitialCommittee` and `getCommitteeMembers` to Conway ImpTest
 * Implement `ConwayUtxowPredFailure` instances:
   * `Arbitrary`
   * `ToExpr`